refactor: vms for sections + fix db manager
