+++ DepNavIOS/Views/ControlSections/SettingsSectionView.swift
@@ -9,8 +9,7 @@ import SwiftUI
 struct SettingsSectionView: View {
     @ObservedObject var mapViewModel: MapViewModel
     @Binding var currentSheetContent: SheetContent
-
-    @ObservedObject var languageManager = LanguageManagerModel.shared
+    @StateObject private var vm: SettingsSectionVM
 
     var body: some View {
         VStack(alignment: .leading, spacing: 16) {
@@ -44,15 +43,14 @@ struct SettingsSectionView: View {
             ForEach(LanguageModel.allCases) { lang in
                 Button(action: {
                     withAnimation {
-                        languageManager.setLanguage(lang)
-                        Bundle.setLanguage(lang.localeIdentifier)
+                        vm.setLanguage(lang)
                     }
                 }) {
                     Text(lang.displayName)
-                        .fontWeight(languageManager.currentLanguage == lang ? .bold : .regular)
-                        .foregroundColor(languageManager.currentLanguage == lang ? .blue : .primary)
+                        .fontWeight(vm.selectedLanguage == lang ? .bold : .regular)
+                        .foregroundColor(vm.selectedLanguage == lang ? .blue : .primary)
                         .padding(6)
-                        .background(languageManager.currentLanguage == lang ? Color.blue.opacity(0.1) : Color.clear)
+                        .background(vm.selectedLanguage == lang ? Color.blue.opacity(0.1) : Color.clear)
                         .cornerRadius(8)
                 }
             }
@@ -71,11 +69,11 @@ struct SettingsSectionView: View {
             Spacer()
 
             ForEach(departments, id: \.id) { dep in
-                let isSelected = mapViewModel.selectedDepartment == dep.id
+                let isSelected = vm.selectedDepartment == dep.id
 
                 Button(action: {
                     withAnimation {
-                        mapViewModel.selectedDepartment = dep.id
+                        vm.setDepartment(dep.id)
                     }
                 }) {
                     Text(dep.display)
@@ -94,5 +92,13 @@ struct SettingsSectionView: View {
         Text(LocalizedString("faq_section_made_by", comment: "Made with love by @qrutyy"))
             .font(.subheadline)
             .foregroundColor(.secondary)
+    }
+}
+
+extension SettingsSectionView {
+    init(mapViewModel: MapViewModel, currentSheetContent: Binding<SheetContent>) {
+        self._mapViewModel = ObservedObject(wrappedValue: mapViewModel)
+        self._currentSheetContent = currentSheetContent
+        _vm = StateObject(wrappedValue: SettingsSectionVM(mapViewModel: mapViewModel))
     }
 }
